split DXF functions in to read / write classes
